backend/db_service/image_net_1000/requirements.txt
Flask == 2.2.2
mysql-connector-python
Pillow
numpy
requests
pandas
grad-cam
